backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
anthropic==0.25.0
python-dotenv==1.0.1
sse-starlette==2.0.0
httpx==0.27.0
